requirements.txt
django
django-cors-headers
djangorestframework
requests
gunicorn
psycopg2-binary